remove staking querier
